# openapi.yaml
openapi: 3.1.0
info:
  title: Notifier
  description: Internal alerting service for Watchdog
  version: 0.0.3
paths:
...
  /internal/v1/api/alertmanager/channels:
    get:
      tags:
      - alertmanager-channels
      summary: List Notification Channels
      description: Lists notification channels visible to the current user.
      operationId: list_channels
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: show_hidden
        in: query
        required: false
        schema:
          type: string
          pattern: ^(true|false)$
          default: 'false'
          title: Show Hidden
      responses:
        '200':
          description: The notification channels visible to the current caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationChannel'
                title: Response List Channels
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - alertmanager-channels
      summary: Create Notification Channel
      description: Creates a new notification channel for the current tenant scope.
      operationId: create_channel
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationChannelCreate'
      responses:
        '201':
          description: The newly created notification channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationChannel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /internal/v1/api/alertmanager/channels/{channel_id}:
    get:
      tags:
      - alertmanager-channels
      summary: Get Notification Channel
      description: Returns a single notification channel when it exists and is visible
        to the current user.
      operationId: get_channel
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: The requested notification channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationChannel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '404':
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - alertmanager-channels
      summary: Update Notification Channel
      description: Updates an existing notification channel in the current tenant
        scope.
      operationId: update_channel
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationChannelCreate'
      responses:
        '200':
          description: The updated notification channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationChannel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - alertmanager-channels
      summary: Delete Notification Channel
      description: Deletes an existing notification channel when the caller has access.
      operationId: delete_channel
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: The deletion result for the notification channel.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Channel
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '404':
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Bad Request
  /internal/v1/api/alertmanager/channels/{channel_id}/hide:
    post:
      tags:
      - alertmanager-channels
      summary: Hide Notification Channel
      description: Toggles whether a shared notification channel is hidden for the
        current user.
      operationId: hide_channel
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/routers__observability__alerts__shared__HideTogglePayload'
      responses:
        '200':
          description: The hide state applied to the notification channel.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Hide Channel
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /internal/v1/api/alertmanager/channels/{channel_id}/test:
    post:
      tags:
      - alertmanager-channels
      summary: Test Notification Channel
      description: Sends a test notification through the specified notification channel.
      operationId: test_channel
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          title: Channel Id
      responses:
        '200':
          description: The test delivery result for the notification channel.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Test Channel
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChannelType:
      type: string
      enum:
      - email
      - slack
      - teams
      - webhook
      - pagerduty
      title: ChannelType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NotificationChannel:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
          description: Unique identifier
          examples:
          - channel-123
        name:
          type: string
          title: Name
          description: Channel name
          examples:
          - Primary Slack
        type:
          $ref: '#/components/schemas/ChannelType'
          description: Channel type
          examples:
          - slack
        enabled:
          type: boolean
          title: Enabled
          description: Whether the channel is enabled
          default: true
          examples:
          - true
        config:
          additionalProperties: true
          type: object
          title: Config
          description: Channel-specific configuration
          examples:
          - webhookUrl: https://hooks.slack.com/services/T000/B000/XXX
        createdBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdby
          description: Owner user id
          examples:
          - user-42
        visibility:
          $ref: '#/components/schemas/Visibility'
          description: Visibility scope
          default: private
          examples:
          - private
        sharedGroupIds:
          items:
            type: string
          type: array
          title: Sharedgroupids
          description: Group IDs this channel is shared with (when visibility=group)
          examples:
          - - group-ops
        isHidden:
          type: boolean
          title: Ishidden
          description: Whether this channel is hidden for the current user
          default: false
          examples:
          - false
      type: object
      required:
      - name
      - type
      - config
      title: NotificationChannel
    NotificationChannelCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
          description: Channel name
          examples:
          - Primary Slack
        type:
          $ref: '#/components/schemas/ChannelType'
          description: Channel type
          examples:
          - slack
        enabled:
          type: boolean
          title: Enabled
          description: Whether the channel is enabled
          default: true
          examples:
          - true
        config:
          additionalProperties: true
          type: object
          title: Config
          description: Channel-specific configuration
          examples:
          - webhookUrl: https://hooks.slack.com/services/T000/B000/XXX
        visibility:
          $ref: '#/components/schemas/Visibility'
          description: Visibility scope
          default: private
          examples:
          - private
        sharedGroupIds:
          items:
            type: string
          type: array
          title: Sharedgroupids
          description: Group IDs to share with
          examples:
          - - group-ops
      additionalProperties: false
      type: object
      required:
      - name
      - type
      - config
      title: NotificationChannelCreate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Visibility:
      type: string
      enum:
      - private
      - group
      - tenant
      - public
      title: Visibility
    routers__observability__alerts__shared__HideTogglePayload:
      properties:
        hidden:
          type: boolean
          title: Hidden
          default: true
          examples:
          - true
      type: object
      title: HideTogglePayload
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...